create `evaluator.py` to separate evaluation workflow from CLI logic
